SelectTree: Move Option into it's own component
